Update phase1.c
This is now calling the test00's XXp1 function, but when it quits, it's not returning a -3 through the system -- maybe your quit fixes will address this?  It's also still not unraveling correctly, though -- instead of start1(): exit status for child 3 is -3
I'm seeing a SIMULATOR TRAP.  That could mean there's still a problem in the join().  That start of join() is definitely working, but there could be a problem in the finish